Fixes and refactor
