+++ activiti-cloud-acceptance-tests-playwright/tests/runtime/process-definition-admin-actions.spec.ts
@@ -15,25 +15,9 @@
  */
 
 import { activiti, expect } from '../../fixtures/services.fixture';
-import { CloudProcessDefinition } from '../../models/process-definition.models';
+import { pickHighestVersionByKey } from '../../helpers/process-definition';
 
 const SINGLE_TASK_PROCESS = 'SingleTaskProcess';
-
-// Mirrors ProcessDefinitionActions#getProcessDefinition — picks the highest appVersion match for a key.
-function findProcessDefinitionByKey(
-    definitions: CloudProcessDefinition[],
-    key: string
-): CloudProcessDefinition {
-    const matches = definitions.filter((def) => def.key === key);
-    if (matches.length === 0) {
-        throw new Error(`No process definition found matching key ${key}`);
-    }
-    return matches.reduce((best, current) => {
-        const bestVersion = parseInt(String(best.appVersion ?? '0'), 10);
-        const currentVersion = parseInt(String(current.appVersion ?? '0'), 10);
-        return currentVersion > bestVersion ? current : best;
-    });
-}
 
 activiti.describe('Process Definition Admin Actions', () => {
     activiti('as an admin I should be able to get process model', async ({ queryAdminServiceHradmin }) => {
@@ -41,7 +25,7 @@ activiti.describe('Process Definition Admin Actions', () => {
             `Then the user, using the admin endpoint, can get the process model for process with key ${SINGLE_TASK_PROCESS} by passing its id`,
             async () => {
                 const definitions = await queryAdminServiceHradmin.getAllProcessDefinitionsAdmin();
-                const definition = findProcessDefinitionByKey(definitions, SINGLE_TASK_PROCESS);
+                const definition = pickHighestVersionByKey(definitions, SINGLE_TASK_PROCESS);
 
                 const processModel = await queryAdminServiceHradmin.getProcessModel(definition.id);
